Move draw related code to D2HDDDraw.
