+++ HD/D2HD.vcxproj
@@ -23,12 +23,14 @@
     <ClInclude Include="..\TemplateIncludes.h" />
     <ClInclude Include="D2HDConfig.h" />
     <ClInclude Include="D2HDPatches.h" />
+    <ClInclude Include="D2HDDraw.h" />
     <ClInclude Include="D2HDStructs.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\DLLmain.cpp" />
     <ClCompile Include="D2HDConfig.cpp" />
     <ClCompile Include="D2HDPatches.cpp" />
+    <ClCompile Include="D2HDDraw.cpp" />
     <ClCompile Include="D2HDStructs.cpp" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
