+++ HD/D2HDDraw.h
@@ -1,0 +1,5 @@
+namespace HD {
+    void RedrawUILeftPanelBorders_Interception();
+    void RedrawUIRightPanelBorders_Interception();
+    
+}
