+++ HD/D2HDPatches.h
@@ -30,8 +30,6 @@ namespace HD {
     int SetupGlideRenderResolution_Interception();
 
     void PanelPosition_Interception();
-    void RedrawUILeftPanelBorders_Interception();
-    void RedrawUIRightPanelBorders_Interception();
     void DrawUIPanelBackground();
 
     int __fastcall GetResolutionMode_Interception();
